% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abolished_mun_key.R
\docType{data}
\name{abolished_mun_key}
\alias{abolished_mun_key}
\title{A key to recode codes of abolished municipalities to corresponding
existing municipalities.}
\format{
A data frame with 660 rows and 2 variables:
\describe{
\item{joiner}{Code of abolished municipality}
\item{joinee}{Code of joinee municipality}
...
}
}
\source{
\% \url{"https://www.stat.fi/static/media/uploads/meta/luokitukset/lakkautetut_kunnat_aakkosjarj20_taul6.xlsx"}
}
\usage{
abolished_mun_key
}
\description{
A key to recode codes of abolished municipalities to corresponding
existing municipalities.
}
\keyword{datasets}
