% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{add_region}
\alias{add_region}
\title{Add regions to data}
\usage{
add_region(data, ..., from = NULL, year = NULL, offline = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{from}{character, variable in the data that is used to add regions to the data. Defaults
to NULL, in which case function tries to guess the variable in the data.}

\item{year, }{double, year of region classification}

\item{offline, }{logical, whether works offline with package data. Defaults to TRUE.}

\item{to}{character, the region to add.}
}
\value{

}
\description{
Uses get_regionkey- and recode-functions to provide a fast and simple way to
add regions to data. Very pipe-friendly.
}
\examples{

}
