% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_region_classifications.R
\name{check_region_name_code_correspondence}
\alias{check_region_name_code_correspondence}
\title{Check if region names and codes correspond as in regionkey.}
\usage{
check_region_name_code_correspondence(
  region_names,
  region_codes,
  year = NULL,
  lang = "fi",
  offline = TRUE
)
}
\arguments{
\item{region_names}{character, vector of region names.}

\item{region_codes}{character, vector region codes.}

\item{year}{year of classification used.}

\item{lang}{language of classification used.}

\item{offline}{logical, whether works offline with package data. Defaults to \code{TRUE}.}
}
\value{
logical
}
\description{
Check if region names and codes correspond as in regionkey.
}
