% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_checking_classifications.R
\name{check_region_names}
\alias{check_region_names}
\alias{check_region_names_vct}
\alias{check_region_names_fct}
\title{Check region names}
\usage{
check_region_names(x, year = NULL, offline = TRUE)

check_region_names_vct(x, year = NULL, offline = TRUE)

check_region_names_fct(x, year = NULL, offline = TRUE)
}
\arguments{
\item{x}{character vector or factor}

\item{year}{double or character, year of classification}

\item{offline}{logical, whether works offline with package data. Defaults to TRUE.}
}
\value{

}
\description{
Check region names
}
\section{Functions}{
\itemize{
\item \code{check_region_names_vct}: Check region names of a vector

For internal use.

\item \code{check_region_names_fct}: Check region names of a factor

For internal use.
}}

