% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_checking_region_classifications.R
\name{check_region_var_name_code_correspondence}
\alias{check_region_var_name_code_correspondence}
\title{Check if region names and codes correspond as in regionkey.}
\usage{
check_region_var_name_code_correspondence(
  data,
  region_name_var,
  region_code_var,
  offline = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{region_name_var}{character the name of the regional name variable in the data}

\item{region_code_var}{character the name of the regional code variable in the data}

\item{offline}{logical, whether works offline with package data. Defaults to TRUE.}
}
\value{
logical
}
\description{
Check if region names and codes correspond as in regionkey.
}
