% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_to_key.R
\name{classification_to_key}
\alias{classification_to_key}
\title{Transform classifications to keys}
\usage{
classification_to_key(.df, level_nchars = NULL)
}
\arguments{
\item{level_nchars}{}
}
\description{
Transform classifications to keys
}
