% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_names.R
\name{codes_to_names}
\alias{codes_to_names}
\title{Changes region codes to region names}
\usage{
codes_to_names(data, region, year)
}
\arguments{
\item{data}{data.frame}

\item{region}{character}

\item{year}{integer}
}
\value{

}
\description{
Changes region codes to region names
}
\examples{

}
