% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{codes_to_names2_df}
\alias{codes_to_names2_df}
\title{Changes region codes to region names}
\usage{
codes_to_names2_df(data, from = NULL, year = NULL)
}
\arguments{
\item{data}{data.frame, the input data that contains a variable of region codes.}

\item{year}{integer, the year of the applied classification key.}

\item{region}{character, the name of the variable of region codes.}
}
\value{
data.frame
}
\description{
A wrapper that uses the statficlassifications::recode-function.
}
\examples{

    # Generate random municipal data
       data <- get_regionkey(year = 2020) \%>\%
               dplyr::select(kunta_code) \%>\%
               mutate(values = rnorm(n()))

       codes_to_names(data)

}
