% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{codes_to_names_vct}
\alias{codes_to_names_vct}
\title{Change codes to names in vector}
\usage{
codes_to_names_vct(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Change codes to names in vector
}
\examples{

  v <- c("SSS", "KU103", "KU061","SK213", "MK04")
  codes_to_names_vct(v)

}
