% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{detect_region_var}
\alias{detect_region_var}
\title{Detect a region variable in data}
\usage{
detect_region_var(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
character(2) Returns the region variable in the original data and the correponding region
varible in the get_regionkey()
}
\description{
Given data, looks for the variable that contains regions. Returns the name of this variable
and the corresponding variable name in get_regionkey()
}
\examples{

   # Generate random municipal data with random name for the regions
             data <- get_regionkey() \%>\%
                     dplyr::select(kunta_name) \%>\%
                     dplyr::rename_with(~paste(sample(letters, 4), collapse = "")) \%>\%
                     dplyr::mutate(values = rnorm(n()))

             detect_region_var(data)

}
