% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_region_variables.R
\name{detect_region_year}
\alias{detect_region_year}
\title{Detect year of region classification}
\usage{
detect_region_year(x, region_level)
}
\arguments{
\item{x}{vector of regions}

\item{region_level}{region level in x}
}
\value{
vector
}
\description{
Detects year of classification from the number of unique regions assuming the
regularity that the number of regions weakly decreases in time. Given a vector of
region names or codes and region level determines the year of regional
classification. Returns all matching years. Works only for a vector that
contains all possible values. Note that if the data does not contain all possible
values the levels of a factor may.
}
\examples{
   df <- get_regionclassification("kunta", year = 2010)
   detect_region_year(df$kunta_name, region_level = "kunta")

}
