% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api.R
\name{filter_results}
\alias{filter_results}
\title{Filter search results}
\usage{
filter_results(results, searchterms)
}
\arguments{
\item{results}{character vector, the search field}

\item{searchterms}{character vector search terms}
}
\value{
character vector of elements of input that match the search terms
}
\description{
Filter search results
}
