% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{find_classification_service}
\alias{find_classification_service}
\title{Finds a classification service for a localId}
\usage{
find_classification_service(localId)
}
\arguments{
\item{localId}{character}
}
\value{

}
\description{
Given a localId finds whether the localId is a localId in classifications or correspondenceTables
classification services. For internal use.
}
\examples{

    find_classification_service(search_series(as_localId = TRUE)[1])

}
