% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_encoding.R
\name{fix_encoding}
\alias{fix_encoding}
\title{Fixes encoding errors in a data frame}
\usage{
fix_encoding(data)
}
\arguments{
\item{data}{a data.frame containing encoding errors.}
}
\value{
data.frame Returns a data.frame where the encoding errors in character variables are fixed.
}
\description{
When importing Statistics Finland classifications, can be used to fix the potential encoding errors.
}
\section{Warning}{

Upper case Ö is missing from the encoding key.
}

\keyword{NA-value}
