% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regionkey.R
\name{get_full_region_code_name_key}
\alias{get_full_region_code_name_key}
\title{Get all region name-code keys in one object.}
\usage{
get_full_region_code_name_key(
  year = NULL,
  offline = TRUE,
  as_named_vector = FALSE
)
}
\arguments{
\item{as_named_vector}{logical, whether returns key as a named vector. Defaults to FALSE,
in which case returns as a data.frame.}
}
\value{

}
\description{
Get all region name-code keys in one object.
}
