% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_key}
\alias{get_key}
\title{Get data from API}
\usage{
get_key(localID, print_key_name = TRUE)
}
\arguments{
\item{localID, }{character, local ID of the required correspondence table}
}
\value{
data.frame
}
\description{
A wrapper for \code{access_API} to get data.
}
