% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_api.R
\name{get_latest_year}
\alias{get_latest_year}
\title{Get the year of the latest correspondence table}
\usage{
get_latest_year(offline = TRUE)
}
\arguments{
\item{offline}{logical, whether works offline with package data. Defaults to TRUE.}
}
\value{
double, the year of the newest correspondence table
}
\description{
Get the year of the latest correspondence table
}
\examples{

 get_latest_year()

}
