% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Get map}
\usage{
get_map(region_level, year)
}
\arguments{
\item{region_level}{character, region level of the map}

\item{year}{double, year of the map}
}
\value{
sf object
}
\description{
Get map
}
