% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occupationkey.R
\name{get_occupationkey}
\alias{get_occupationkey}
\title{Get an occupation name-code key}
\usage{
get_occupationkey(code_level = NULL, year = 2010)
}
\arguments{
\item{code_level, }{disaggregation level}

\item{year, }{year of the key}
}
\value{
data.frame with occupation codes and names
}
\description{
Get an occupation name-code key
}
\examples{

   get_occupationkey(1)

}
