% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_get_region_info.R
\name{get_regionclassification}
\alias{get_regionclassification}
\title{Get region classifications / code-name keys}
\usage{
get_regionclassification(
  ...,
  year = NULL,
  offline = TRUE,
  as_named_vector = FALSE,
  suppress_message = FALSE,
  only_names = FALSE,
  only_codes = FALSE
)
}
\arguments{
\item{...}{character(s), (vector), region(s) of required keys.}

\item{year}{character/numeric, year of the required keys. If NULL uses the latest year.}

\item{offline}{logical, whether uses the key in the package data. Defaults TRUE.}

\item{as_named_vector}{logical, whether returns the key as a named vector rather than a
data.frame. Defaults FALSE.}

\item{suppress_message, }{logical, whether to suppress any messages the function might produce.}

\item{only_names, }{logical, whether to return only the names in the classification.
Defaults to FALSE.}

\item{only_codes, logical, }{whether to return only the codes in the classification.
Defaults to FALSE.}
}
\value{
a data.frame or a named vector. A region code-name key.
}
\description{
Get region classifications / code-name keys
}
\examples{

get_regionclassification("seutukunta")
get_regionclassification("seutukunta", as_named_vector = TRUE)
get_regionclassification("seutukunta", "maakunta")

}
