% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regionkey.R
\name{get_regionkey}
\alias{get_regionkey}
\title{Region classification keys}
\usage{
get_regionkey(region = "kunta", only_codes = FALSE, only_names = FALSE, year)
}
\arguments{
\item{region}{character, the smallest region desired in the resulting classification key.}

\item{year}{character or numerical, the year of desired classification key.}

\item{codes}{logical, whether the classification key contains region names or region codes. Defaults to FALSE.}
}
\value{
data.frame Returns a classification key as a data.frame.
}
\description{
Imports NUTS-region classification keys from Statistics Finland. Use together with 'dplyr::left_join'
to add regions to data.
}
\examples{

}
