% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_series}
\alias{get_series}
\title{Get classification series from API}
\usage{
get_series(localId, print_series_name = TRUE)
}
\arguments{
\item{localId}{character, local ID of the required correspondence table}

\item{print_series_name}{whether prints the long name of the classification series.}
}
\value{

}
\description{
A wrapper for \code{access_API} to get data of classification series
}
\examples{

  localId <- "siviiliasiat_1_20140101"
  get_series(localId)

}
