% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_checking_classifications.R
\name{is_region_code}
\alias{is_region_code}
\alias{is_region_code_with_prefix}
\alias{is_region_code_without_prefix}
\title{Check if character is prefixed region code}
\usage{
is_region_code(x, region_level = NULL, year = NULL, offline = TRUE)

is_region_code_with_prefix(x, region_level = NULL, year = NULL, offline = TRUE)

is_region_code_without_prefix(
  x,
  region_level = NULL,
  year = NULL,
  offline = TRUE
)
}
\arguments{
\item{x}{character}

\item{region_level}{character, optional region level of the input region codes}

\item{year}{double}

\item{offline}{logical, whether works offline with package data. Defaults to TRUE.}
}
\value{

}
\description{
Check if character is prefixed region code
}
\section{Functions}{
\itemize{
\item \code{is_region_code_with_prefix}: Check if input is region code with prefix.

\item \code{is_region_code_without_prefix}: Check if input is region code without prefix.
}}

\examples{

  is_region_code(c("KU005", "005"))
  is_region_code_with_prefix(c("KU005", "005"))
  is_region_code_without_prefix(c("KU005", "005"))
  is_region_code_with_prefix(c("KU005", "005"), region_level = "maakunta")
  is_region_code_without_prefix(c("KU005", "005"), region_level = "maakunta")

}
