% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_region_classifications.R
\name{is_region_name}
\alias{is_region_name}
\title{Check if string is region name}
\usage{
is_region_name(
  x,
  region_level = NULL,
  year = NULL,
  offline = TRUE,
  allow_nonstandard_names = FALSE,
  case_sensitive = TRUE,
  lang = "fi"
)
}
\arguments{
\item{x}{character, a vector of potential region names.}

\item{region_level}{character, optional region level of the input region codes.
Possible \code{region_level}s are \code{statficlassifications:::prefix_name_key$name}.}

\item{year}{double, year of region classification searched.}

\item{offline}{logical, whether works offline with package data. Defaults to \code{TRUE}.}

\item{allow_nonstandard_names}{logical, whether to accept a broader set of names as
region names.}

\item{case_sensitive}{logical, whether recognition is case sensitive, defaults
to \code{FALSE}.}

\item{lang}{\code{"fi"}, \code{"sv"} or \code{"en"}. Language of the input name.
Defaults to \code{"fi"}.}
}
\value{
logical
}
\description{
By default, the tests whether input is a name in standard form. Non-standard
forms can be tested by setting \code{allow_nonstandard_names = TRUE}.
}
\details{
Region variable detection functions use the testing functions here.
}
\examples{

 is_region_name(c("Kainuu", "Kainuun maakunta", "Kajjjaani"))
 is_region_name(c("Kainuu", "Kainuun maakunta", "Kajjjaani"), allow_nonstandard_names = TRUE)


}
