% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_abolished_mun.R
\name{join_abolished_mun}
\alias{join_abolished_mun}
\title{Join abolished municipalities}
\usage{
join_abolished_mun(x, col = NULL)
}
\arguments{
\item{x, }{an object containing municipality codes, a vector, a factor or a data.frame.}
}
\value{

}
\description{
Takes a vector of municipality codes with KU-prefixes and transforms the codes of
the abolished municipalities to the codes of the municipality to which they have
joined.
}
\examples{

  v <- c("KU414", "KU609", "KU429", "KU272")
  join_abolished_mun(v)
  f <- factor(c("KU414", "KU609", "KU429", "KU272"))
  join_abolished_mun(f)
  df <- data.frame(kunta_code =  c("KU414", "KU609", "KU429", "KU272"), values = rnorm(4))
  join_abolished_mun(df, "kunta_code")
  df <- data.frame(kunta_code =  factor(c("KU414", "KU609", "KU429", "KU272")), values = rnorm(4))
  join_abolished_mun(df, "kunta_code")

}
