% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{join_abolished_municipalities}
\alias{join_abolished_municipalities}
\title{Join abolished municipalities}
\usage{
join_abolished_municipalities(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Join abolished municipalities
}
\examples{

  v <- c("KU414", "KU609", "KU429", "KU273")
  join_abolished_municipalities(v)

}
