% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_recode.R
\name{key_recode_internal}
\alias{key_recode_internal}
\title{Recode with list containing key and source and target variable names}
\usage{
key_recode_internal(x, pkey)
}
\arguments{
\item{x}{vector of elements to be recoded.}

\item{pkey}{list, output of \code{prepare_key()}.}
}
\value{
data.frame or vector
}
\description{
For internal use
}
