% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_region_codes.R
\name{match_region_codes_internal}
\alias{match_region_codes_internal}
\title{match_region_codes Match single region code. For internal use.}
\usage{
match_region_codes_internal(
  x,
  key,
  region_level = NULL,
  use_char_length_info = NULL
)
}
\arguments{
\item{x}{input code}

\item{key}{data.frame or named vector, region key}

\item{region_level}{character (vector) region level of the input codes}

\item{use_char_length_info}{TRUE or named vector, whether to use code character length
information in determining their region level. Defaults to \code{NULL}.}
}
\value{
vector
}
\description{
match_region_codes Match single region code. For internal use.
}
