% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_to_codes.R
\name{names_to_codes}
\alias{names_to_codes}
\title{Changes region names to region codes}
\usage{
names_to_codes(data, region, year)
}
\arguments{
\item{data}{data.frame}

\item{region}{character}

\item{year}{integer}
}
\value{

}
\description{
Changes region names to region codes
}
\examples{

}
