% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{names_to_codes2_df}
\alias{names_to_codes2_df}
\title{Change region names to region codes}
\usage{
names_to_codes2_df(data, from = NULL, year = NULL)
}
\arguments{
\item{data}{data.frame, the input data that contains a variable of region names.}

\item{year}{integer, the year of the applied classification key.}

\item{region}{character, the name of the variable of region names.}
}
\value{
data.frame
}
\description{
#' A wrapper that uses the statficlassifications::recode-function.
}
\examples{

#' # Replace municipality names by their codes

    # Generate random municipal data
       data <- get_regionkey(year = 2020) \%>\%
               dplyr::select(kunta_name) \%>\%
               mutate(values = rnorm(n()))

    names_to_codes(data)

}
