% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonstandard_region_names_key.R
\docType{data}
\name{nonstandard_region_names_key}
\alias{nonstandard_region_names_key}
\title{Nonstandard region names key}
\format{
A data frame with ?? rows and ?? variables:
\describe{
\item{}{}
\item{}{}
...
}
}
\source{
\url{}
}
\usage{
nonstandard_region_names_key
}
\description{
Nonstandard region names key
}
\keyword{datasets}
