% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_region_codes.R
\name{numeric_code_to_character}
\alias{numeric_code_to_character}
\title{Title}
\usage{
numeric_code_to_character(x, region_level_prefix)
}
\arguments{
\item{region_level_prefix}{}
}
\value{

}
\description{
Title
}
