% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_current_mun_key.R
\docType{data}
\name{old_current_mun_key}
\alias{old_current_mun_key}
\title{Old current mun key}
\format{
A data frame with 660 rows and 2 variables:
\describe{
\item{}{}
\item{}{}
...
}
}
\source{
\url{}
}
\usage{
old_current_mun_key
}
\description{
Old current mun key
}
\keyword{datasets}
