% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_name_key.R
\docType{data}
\name{prefix_name_key}
\alias{prefix_name_key}
\title{Prefix-name key.}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{}{}
\item{}{}
...
}
}
\source{
\url{}
}
\usage{
prefix_name_key
}
\description{
Prefix-name key.
}
\keyword{datasets}
