% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_names_to_codes.R
\name{prefix_to_name}
\alias{prefix_to_name}
\title{Change region prefixes to names}
\usage{
prefix_to_name(prefix)
}
\arguments{
\item{prefix, }{region level code prefix}
}
\value{
region name
}
\description{
Change region prefixes to names
}
\examples{

 prefix_to_name("SK")

}
