% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_code_to_name_key.R
\docType{data}
\name{region_code_to_name_key}
\alias{region_code_to_name_key}
\title{Region code to name key}
\format{
A data frame with 310 rows and 10 variables:
}
\usage{
region_code_to_name_key
}
\description{
Region code to name key
}
\keyword{datasets}
