% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_name_to_code_key.R
\docType{data}
\name{region_name_to_code_key}
\alias{region_name_to_code_key}
\title{Region name to code key}
\format{
A data frame with 310 rows and 10 variables:
\describe{
\item{}{}
\item{}{}
...
}
}
\source{
\url{}
}
\usage{
region_name_to_code_key
}
\description{
Region name to code key
}
\keyword{datasets}
