% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionkey.R
\docType{data}
\name{regionkey}
\alias{regionkey}
\title{Regionkey}
\format{
A data frame with 310 rows and 8 variables:
\describe{
\item{kunta_name}{}
\item{seutukunta_name}{}
\item{maakunta_name}{}
\item{suuralue_name}{}
\item{ely_name}{}
\item{kunta_code}{}
\item{seutukunta_code}{}
\item{maakunta_code}{}
\item{suuralue_code}{}
\item{ely_code}{}
}
}
\source{
\url{}
}
\usage{
regionkey
}
\description{
A correspondence table for the most used regions names and codes.
}
\keyword{datasets}
