% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_classifications.R
\name{search_classifications}
\alias{search_classifications}
\title{Search for classifications}
\usage{
search_classifications(..., year = NULL, as_localId = FALSE)
}
\arguments{
\item{...}{character, search words.}

\item{year}{character or numerical, search for specific years.}

\item{as_localId}{logical, whether returns the localID of the found table. Defaults to FALSE.}
}
\value{

}
\description{
Search for classifications
}
\examples{

  search_classifications("ammatti")
  search_classifications("ammatti", year = 2021, as_localId = TRUE)

}
