% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api.R
\name{search_keys}
\alias{search_keys}
\title{Search for classification keys}
\usage{
search_keys(..., as_localId = FALSE)
}
\arguments{
\item{...}{character, search words.}

\item{as_localId}{logical, whether returns the localID of the found table. Defaults to \code{FALSE}.}
}
\value{
character vector
}
\description{
Searches and browses available classification keys / correspondence tables in
the correspondenceTables classification service
}
\examples{

# Browse all keys
   search_keys()

# Search for keys with search word "maakunta"
   search_keys("maakunta")

# Search for keys from kunta to maakunta for year 2016
   search_keys("kunta", "maakunta", 2016)

}
