% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_series.R
\name{search_series}
\alias{search_series}
\title{Search for classification series}
\usage{
search_series(..., year = NULL, as_localId = FALSE)
}
\arguments{
\item{...}{character, search words.}

\item{year}{character or numerical, search for specific years.}

\item{as_localId}{logical, whether returns the localID of the found table. Defaults to FALSE.}
}
\value{

}
\description{
Search for classification series
}
\examples{

  search_series("ammatti")
  search_series("ammatti", year = 2021, as_localId = TRUE)

}
