% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{set_region_codes}
\alias{set_region_codes}
\title{Standardize region codes with prefixes}
\usage{
set_region_codes(x)
}
\arguments{
\item{x}{character vector of region codes}
}
\value{

}
\description{
Standardize region codes with prefixes
}
\examples{

  v <- c("020" = "Akaa", "047" = "Enontekiö", "15" = "Pohjanmaa", "133" = "Keuruu")
  set_region_codes(names(v))
  v <- c("020" = "Akaa", "047" = "Enontekiö", "MK15" = "Pohjanmaa", "SK133" = "Keuruu")
  set_region_codes(names(v))
  v <- c("020" = "Akaa", "KU047" = "Enontekiö", "15" = "Pohjanmaa", "133" = "Keuruu")
  set_region_codes(names(v))

}
