% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{standardize_code_prefixes}
\alias{standardize_code_prefixes}
\title{Standardize region codes with prefixes}
\usage{
standardize_code_prefixes(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Standardize region codes with prefixes
}
\examples{

  v <- c("020" = "Akaa", "047" = "Enontekiö", "15" = "Pohjanmaa", "133" = "Keuruu")
  standardize_code_prefixes_vct(names(v))
  v <- c("020" = "Akaa", "047" = "Enontekiö", "MK15" = "Pohjanmaa", "SK133" = "Keuruu")
  standardize_code_prefixes_vct(names(v))

}
