% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statficlassifications-package.R
\docType{package}
\name{statficlassifications-package}
\alias{statficlassifications}
\alias{statficlassifications-package}
\title{statficlassifications: Statistics Finland Statistical Classification API interface}
\description{
statficlassifications accesses to the classifications and correspondence tables open classifications API of Statistics Finland. Contains tools for checking and standardizing classifications.
}
\author{
\strong{Maintainer}: Juho Alasalmi \email{juho.alasalmi@ptt.fi}

Other contributors:
\itemize{
  \item Janne Huovari [contributor]
  \item Pellervo Economic Research PTT [funder]
}

}
\keyword{internal}
