% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statficlassifications_api.R
\name{statficlassifications_api}
\alias{statficlassifications_api}
\title{API interface}
\usage{
statficlassifications_api(localID)
}
\arguments{
\item{localID}{character local ID of the required correspondence table}
}
\value{
data.frame
}
\description{
Given a local ID of a correspondence table fetches the table and reads it.
}
