% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{urls_as_localId_df}
\alias{urls_as_localId_df}
\title{Transform a list of urls into a data.frame that separates the relevant information of each localId.}
\usage{
urls_as_localId_df(urls)
}
\arguments{
\item{urls, }{a list of characters, urls of localIds..}
}
\value{
data.frame containing the relevant information of each localId correponding to the input urls.
}
\description{
Used when searching localIds among endpoints. For internal use.
}
