% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{access_API}
\alias{access_API}
\title{API interface}
\usage{
access_API(localId = NULL, content = "data", classification_service = NULL)
}
\arguments{
\item{content, }{character, either "data" or "url" whether the content of the query
is data or url.}

\item{classification_service}{character, either 'correspondenceTable' or
'classifications'. Determines the classification service used.}

\item{localID}{character, local ID of the correspondence table or classification}
}
\value{
data.frame either the correspondence table or its url depending on argument \code{content}.
}
\description{
Given a local ID of a correspondence table or classification, gets the data
and transforms it into a data.frame. For internal use of the package.
}
\details{
Currently supported classification services are 'classifications' and
'correspondenceTables'. For more information on classification services,
see \url{https://www.stat.fi/en/luokitukset/info/}.
}
\examples{

  localId <- "kunta_1_20200101\%23seutukunta_1_20200101"
  access_API(localId, content = "data")
  access_API(localId, content = "url")

}
