% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{access_API}
\alias{access_API}
\title{API interface}
\usage{
access_API(localID = NULL, content = "data")
}
\arguments{
\item{localID}{character local ID of the required correspondence table}

\item{content, }{character either "data" or "url" whether the content of the query
is data or url.}
}
\value{
data.frame
}
\description{
Given a local ID of a correspondence table fetches the table and reads it.
}
