% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{add_region}
\alias{add_region}
\title{Add regions to data}
\usage{
add_region(data, to, from = NULL)
}
\arguments{
\item{data}{data.frame}

\item{to}{character, the region to add.}

\item{from}{character, variable in the data that is used to add regions to the data. Defaults
to NULL, in which case function tries to guess the variable in the data.}
}
\value{

}
\description{
Uses \code(get_regionkey)- and \code(recode)-functions to provide a fast and simple way to
add regions to data. Very pipe-friendly.
}
\examples{

}
