% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_checking_region_classifications.R
\name{check_region_codes}
\alias{check_region_codes}
\alias{check_region_codes_vct}
\alias{check_region_codes_fct}
\title{Check region codes}
\usage{
check_region_codes(x, year = NULL, offline = TRUE)

check_region_codes_vct(x, year = NULL, offline = TRUE)

check_region_codes_fct(x, year = NULL, offline = TRUE)
}
\arguments{
\item{x}{character vector or factor}

\item{year}{double or character, year of classification}

\item{offline}{logical, whether works offline with package data. Defaults to TRUE.}
}
\value{
vector of logicals
}
\description{
Checks whether all the elements in the given vector of potential region codes
are standardized region codes. If they are returns \code{TRUE}, if they are
not, returns a message telling which elements are not recognized as region codes.
}
\section{Functions}{
\itemize{
\item \code{check_region_codes_vct}: Check region codes of a vector

For internal use.

\item \code{check_region_codes_fct}: Check region codes of a vector

For internal use.
}}

