% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_region_classifications.R
\name{check_region_name_code_correspondence}
\alias{check_region_name_code_correspondence}
\title{Check if region names and codes correspond as in regionkey.}
\usage{
check_region_name_code_correspondence(
  region_names,
  region_codes,
  year = NULL,
  lang = "fi",
  offline = TRUE
)
}
\arguments{
\item{offline}{logical, whether works offline with package data. Defaults to TRUE.}

\item{data}{data.frame}

\item{region_name_var}{character the name of the regional name variable in the data}

\item{region_code_var}{character the name of the regional code variable in the data}
}
\value{
logical
}
\description{
Check if region names and codes correspond as in regionkey.
}
