% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_checking_classifications.R
\name{check_region_var_classification}
\alias{check_region_var_classification}
\title{Check if a region variable correspond to classification of regionkey}
\usage{
check_region_var_classification(data, region_var, offline = TRUE)
}
\arguments{
\item{offline}{}
}
\value{

}
\description{
Check if a region variable correspond to classification of regionkey
}
