% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_checking_classifications.R
\name{check_region_var_name_code_correspondence}
\alias{check_region_var_name_code_correspondence}
\title{Check if region names and codes correspond as in regionkey.}
\usage{
check_region_var_name_code_correspondence(
  data,
  region_name_var,
  region_code_var,
  offline = TRUE
)
}
\arguments{
\item{offline}{}
}
\value{

}
\description{
Check if region names and codes correspond as in regionkey.
}
