% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_names_to_codes.R
\name{codes_to_names}
\alias{codes_to_names}
\title{Changes region codes to region names}
\usage{
codes_to_names(x, col = NULL)
}
\arguments{
\item{data}{data.frame, the input data that contains a variable of region codes.}

\item{region}{character, the name of the variable of region codes.}

\item{year}{integer, the year of the applied classification key.}
}
\value{
data.frame
}
\description{
A wrapper that uses the statficlassifications::recode-function.
}
\examples{

  v <- c("SSS", "KU103", "KU061","SK213", "MK04")
  codes_to_names(v)
  f <- factor(c("SSS", "KU103", "KU061","SK213", "MK04"))
  codes_to_names(f)
  df <- data.frame(kunta_code = c("SSS", "KU103", "KU061","SK213", "MK04"), values = rnorm(5))
  codes_to_names(df, "kunta_code")


}
