% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_deal_with_region_codes_and_names.R
\name{codes_to_names}
\alias{codes_to_names}
\alias{codes_to_names_vct}
\alias{codes_to_names_fct}
\title{Change region codes to region names}
\usage{
codes_to_names(
  x,
  region_level = NULL,
  use_char_length_info = NULL,
  year = NULL,
  lang = "fi",
  offline = TRUE,
  region_codes_check = FALSE
)

codes_to_names_vct(x, year = NULL, lang = "fi", offline = TRUE)

codes_to_names_fct(x, year = NULL, lang = "fi", offline = TRUE)
}
\arguments{
\item{x}{a character (vector) of region codes}

\item{region_level}{character, optional region level of the input region codes}

\item{use_char_length_info}{TRUE or named vector, whether to use code character length
information in determining their region level. Defaults to NULL.}

\item{year}{integer, the year of the applied classification key.}

\item{lang}{\code{"fi"}, \code{"sv"}, \code{"en"}. Language of output names.
Defaults to \code{"fi"}.}

\item{offline}{logical, whether works offline with package data. Defaults to \code{TRUE}.}

\item{region_codes_check}{logical, whether tries to set standard region codes.
Defaults to \code{FALSE}.}
}
\value{
data.frame
}
\description{
Change region codes to region names
}
\section{Functions}{
\itemize{
\item \code{codes_to_names_vct}: Change region codes to region names

For internal use

\item \code{codes_to_names_fct}: Change region codes to region names

For internal use.
}}

\examples{

  v <- c("SSS", "KU103", "KU061","SK213", "MK04")
  codes_to_names(v)
  v <- c("SSS", "KU103", "KU061","SK213", "MK04", "nav9w4t")
  codes_to_names(v)
  f <- factor(c("SSS", "KU103", "KU061","SK213", "MK04"))
  codes_to_names(f)

}
