% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_api.R
\name{create_localId_name}
\alias{create_localId_name}
\title{Create correspondence table localId}
\usage{
create_localId_name(
  source,
  target,
  year,
  year1 = year,
  year2 = year,
  date = "0101",
  date1 = date,
  date2 = date,
  nro1 = "_1_",
  nro2 = "_1_",
  input_vector = NULL
)
}
\arguments{
\item{source}{character}

\item{target}{character}

\item{year}{integer}

\item{year1}{integer}

\item{year2}{integer}

\item{date}{character defaults to "0101" which the most correspondence tables have.}

\item{date1}{character}

\item{date2}{character}

\item{nro1}{character, defaults to "\emph{1}".}

\item{nro2}{character, defaults to "\emph{1}".}

\item{input_vector}{named vector}
}
\value{
character
}
\description{
Given the inputs, creates a localId. For internal use.
}
\examples{

# Create a localId for the key that maps "kunta" to "maakunta" for year 2015
   create_localId_name("kunta", "maakunta", year = 2015)

}
