% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_localID_name.R
\name{create_localID_name}
\alias{create_localID_name}
\title{Create correspondence table local ID}
\usage{
create_localID_name(source, target, year, date = "0101")
}
\arguments{
\item{source}{character}

\item{target}{character}

\item{year}{integer}

\item{date}{character defaults to "0101" which the most correspondence tables have.}
}
\value{
character
}
\description{
Given the inputs creates the local ID of a correspondence table. For internal use.
}
\examples{

}
