% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_region_variables.R
\name{detect_region_var}
\alias{detect_region_var}
\title{Detect region variables in data}
\usage{
detect_region_var(data, year = NULL, offline = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{year, }{double, year of the used classification}

\item{offline, }{logical, whether works offline with package data. Defaults to \code{TRUE}.}
}
\value{
character(2) Returns the region variable in the original data and the corresponding region
variable in the \code{get_regionkey()}
}
\description{
Given data, looks for the variable that contains regions. Returns the name of this variable
and the corresponding variable name in get_regionkey()
}
\examples{

   # Generate random municipal data with random name for the regions
             data <- get_regionkey() |>
                     dplyr::select(kunta_name) |>
                     dplyr::rename_with(~paste(sample(letters, 4), collapse = "")) |>
                     dplyr::mutate(values = rnorm(dplyr::n()))

             detect_region_var(data)

}
