% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_region_variable_detection.R
\name{detect_region_year}
\alias{detect_region_year}
\title{Detect year of region classification}
\usage{
detect_region_year(x, region_level)
}
\arguments{
\item{region_level}{}
}
\value{
vector
}
\description{
Detects year of classification from the number of unique regions using the idea that
the number of regions vary yearly. Given a vector of region names or codes and region
level determines the year of regional classification. Returns all matching years.
}
\examples{
   df <- get_regionclassification("kunta", year = 2010)
   detect_region_year(df$kunta_name, region_level = "kunta")

}
