% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_api.R
\name{find_classification_service}
\alias{find_classification_service}
\title{Finds a classification service for a localId}
\usage{
find_classification_service(localId)
}
\arguments{
\item{localId}{character}
}
\value{
classification service name(s), character vector
}
\description{
Given a localId finds whether the it is a localId in classifications or correspondenceTables
classification services. For internal use.
}
\examples{

    find_classification_service("siviiliasiat_1_20140101")
    find_classification_service("instit_sektori_1_19960101\%23sektoriluokitus_1_20000101")

}
