% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_classification}
\alias{get_classification}
\title{Get classification series from API}
\usage{
get_classification(localId, print_series_name = TRUE, as_named_vector = FALSE)
}
\arguments{
\item{localId}{character, local ID of the required correspondence table}

\item{print_series_name}{whether prints the long name of the classification series.}

\item{as_named_vector, }{logical, whether to return the object as a named vector rather
than data.frame. Defaults to FALSE.}
}
\value{

}
\description{
A wrapper for \code{access_API} to get data of classifications.
}
\examples{

  localId <- "siviiliasiat_1_20140101"
  get_classification(localId)

}
