% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_key}
\alias{get_key}
\title{API interface}
\usage{
get_key(localID)
}
\arguments{
\item{localID}{character local ID of the required correspondence table}
}
\value{
data.frame
}
\description{
Given a local ID of a correspondence table fetches the table and reads it.
}
