% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_key}
\alias{get_key}
\title{Get classification keys}
\usage{
get_key(localId, lang = "fi")
}
\arguments{
\item{localId}{character, local ID of the required correspondence table}

\item{lang}{\code{"fi"}, \code{"en"}, or \code{"sv"}, desired language.
Defaults to \code{"fi"}.}
}
\value{
data.frame, the key of the provided localId.
}
\description{
A wrapper for \code{access_API} to get classification keys.
}
\examples{

   localId <- "kunta_1_20200101\%23seutukunta_1_20200101"
   get_key(localId)

}
