% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_key}
\alias{get_key}
\title{Get classification key from API}
\usage{
get_key(localId, lang = "fi", print_key_name = TRUE)
}
\arguments{
\item{localId, }{character, local ID of the required correspondence table}

\item{lang, }{\code{fi}, \code{en}, or \code{sv}, language of the key required.
Defaults to \code{fi}.}

\item{print_key_name, }{whether prints the long name of the correspondence table.
Defaults to \code{TRUE}.}
}
\value{
data.frame, the key of the provided localId.
}
\description{
A wrapper for \code{access_API} to get data of classification keys.
}
\examples{

   localId <- "kunta_1_20200101\%23seutukunta_1_20200101"
   get_key(localId)

}
