% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_latest_year}
\alias{get_latest_year}
\title{Get the year of the newest correspondence table}
\usage{
get_latest_year()
}
\value{
double, the year of the newest correspondence table
}
\description{
Get the year of the newest correspondence table
}
