% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regionkey.R
\name{get_region_code_name_key}
\alias{get_region_code_name_key}
\title{Get region name-code keys.}
\usage{
get_region_code_name_key(region, year = NULL, offline = TRUE)
}
\arguments{
\item{region}{character (vector) region(s) of required keys.}

\item{offline}{}
}
\value{
data.frame key
}
\description{
Get region name-code keys.
}
\examples{

  get_region_code_name_key("seutukunta")

}
