% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_classifications.R
\name{get_regionclassification}
\alias{get_regionclassification}
\title{Get region classifications / code-name keys}
\usage{
get_regionclassification(
  ...,
  year = NULL,
  only_names = FALSE,
  only_codes = FALSE,
  as_named_vector = FALSE,
  lang = "fi",
  offline = TRUE
)
}
\arguments{
\item{...}{character(s), (vector), region(s) of required keys.}

\item{year}{character/numeric, year of the required keys. As default returns a general
classification containing also abolished municipalities.}

\item{only_names}{logical, whether to return only the names in the classification.
Defaults to \code{FALSE}.}

\item{only_codes}{logical, whether to return only the codes in the classification.
Defaults to \code{FALSE}.}

\item{as_named_vector}{logical, whether returns the key as a named vector rather than a
data.frame. Defaults \code{FALSE}.}

\item{lang}{\code{"fi"}, \code{"sv"} or \code{"en"}. Language of the classification required.
Defaults to \code{"fi"}.}

\item{offline}{logical, whether uses the key in the package data. Defaults \code{TRUE}.}

\item{suppress_message}{logical, whether to suppress any messages the function might produce.}
}
\value{
a data.frame or a named vector. A region code-name key.
}
\description{
Get region classifications / code-name keys
}
\examples{

get_regionclassification("seutukunta")
get_regionclassification("seutukunta", as_named_vector = TRUE)
get_regionclassification("seutukunta", "maakunta")

}
