% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_classifications.R
\name{get_regionkey}
\alias{get_regionkey}
\title{Get region classification key}
\usage{
get_regionkey(
  ...,
  year = NULL,
  only_codes = FALSE,
  only_names = FALSE,
  lang = "fi",
  offline = TRUE
)
}
\arguments{
\item{...}{character(s), the regions to include in the key.}

\item{year}{character or numerical, the year of the desired classification key.}

\item{only_codes}{logical, whether the key should contain only the region codes.
Defaults to \code{FALSE}.}

\item{only_names}{logical, whether the key should contain only the region names.
Defaults to \code{FALSE}.}

\item{lang}{\code{"fi"}, \code{"sv"} or \code{"en"}. Language of the key required.
Defaults to \code{"fi"}.}

\item{offline}{logical, whether uses the key in the package data. Defaults \code{TRUE}.}
}
\value{
Region classification key as a data.frame.
}
\description{
Imports region classification keys from Statistics Finland API. A wrapper for
\code{get_key()} that it calls under the hood.
}
\details{
The classification keys e.g. from seutukunta to maakunta or maakunta to suuralue are not very well
available. Thus, the logic of the function is to first get all keys from kunta to all other regions,
construct a table that contains all municipalities with their corresponding larger regions and
then apply the selection implied by the arguments set by the user possible removing duplicates.
\code{get_regionkey()} can thus also construct keys that are not available in the Statistics
Finland classification keys API.
}
\examples{

get_regionkey()
get_regionkey("kunta", "seutukunta", only_codes = TRUE)

}
