% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_get_region_info.R
\name{get_regionkey}
\alias{get_regionkey}
\title{Get region classification key}
\usage{
get_regionkey(
  from = "kunta",
  ...,
  year = NULL,
  lang = "fi",
  only_codes = FALSE,
  only_names = FALSE,
  offline = TRUE
)
}
\arguments{
\item{from}{character, the smallest region desired in the resulting classification key.}

\item{...}{character(s), the regions to include in the key other than \code{from}.}

\item{year}{character or numerical, the year of the desired classification key.}

\item{lang}{\code{"fi"}, \code{"en"}, or \code{"sv"}, language of the key required.
Defaults to \code{"fi"}.}

\item{only_codes}{logical, whether the key should contain only the region codes.
Defaults to \code{FALSE}.}

\item{only_names}{logical, whether the key should contain only the region names.
Defaults to \code{FALSE}.}

\item{offline}{logical, whether uses the key in the package data. Defaults \code{TRUE}.}
}
\value{
Region classification key as a data.frame.
}
\description{
Imports region classification keys from Statistics Finland API. Use together with
\code{dplyr::left_join} to add regions to data. A wrapper for the
\code{get_key}-function that it calls under the hood.
}
\details{
The classification keys e.g. from seutukunta to maakunta or maakunta to suuralue are not very well
available. Thus, the logic of the function is to first get all keys from kunta to all other regions,
construct a table that contains all municipalities with their corresponding larger regions and
then apply the selection implied by the arguments set by the user possible removing duplicates.
}
\examples{

get_regionkey()
get_regionkey("kunta", "seutukunta", only_codes = TRUE)

}
