% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_url}
\alias{get_url}
\title{Get urls from API}
\usage{
get_url(localID = NULL)
}
\arguments{
\item{localID, }{character, local ID of the required correspondence table}
}
\value{
character,
}
\description{
A wrapper for \code{access_API} function to get url.
}
