% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_access.R
\name{get_url}
\alias{get_url}
\title{Get url from API}
\usage{
get_url(localId = NULL, classification_service = NULL)
}
\arguments{
\item{localId, }{character, localId of the required table.}
}
\value{
character, url of the provided localId.
}
\description{
A wrapper for \code{access_API} function to get url.
}
\examples{

   localId <- "kunta_1_20200101\%23seutukunta_1_20200101"
   get_url(localId, classification_service = "correspondenceTables")

}
