% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_region_classifications.R
\name{is_region_code}
\alias{is_region_code}
\title{Check if character is prefixed region code}
\usage{
is_region_code(x, region_level = NULL, year = NULL, offline = TRUE)
}
\arguments{
\item{x}{vector, potential region code}

\item{region_level}{character, optional region level of the input region codes}

\item{year}{double, year of region classification searched}

\item{offline}{logical, whether works offline with package data. Defaults to \code{TRUE}.}
}
\value{
logical
}
\description{
Given a vector of potential region codes, returns a vector of logicals indicating
which elements are region codes. Allows prefixed and non-prefixed region codes.
}
\details{
Uses \code{is_region_code_with_prefix()} and \code{is_region_code_without_prefix()}.
These functions are required as such elsewhere.

Region variable detection functions use the testing functions here.
}
\examples{

  is_region_code(c("KU005", "005"))
  is_region_code_with_prefix(c("KU005", "005"))
  is_region_code_without_prefix(c("KU005", "005"))
  is_region_code_with_prefix(c("KU005", "005"), region_level = "maakunta")
  is_region_code_without_prefix(c("KU005", "005"), region_level = "maakunta")

}
