% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_testing_region_classifications.R
\name{is_region_name}
\alias{is_region_name}
\title{Check if character is region name.}
\usage{
is_region_name(
  x,
  region_level = NULL,
  year = NULL,
  offline = TRUE,
  allow_nonstandard_names = FALSE,
  case_sensitive = TRUE,
  lang = "fi"
)
}
\arguments{
\item{x}{character, a vector of potential region names}

\item{region_level}{character, optional region level of the input region codes}

\item{year}{double}

\item{offline}{logical, whether works offline with package data. Defaults to \code{TRUE}.}

\item{allow_nonstandard_names}{logical, whether to accept a broader set of names as
region names.}

\item{case_sensitive, }{logical, whether recognition is case sensitive, defaults
to \code{FALSE}}

\item{lang, }{\code{fi}, \code{sv} or \code{en}. Language of the input name.
Defaults to \code{fi}.}
}
\value{
logical
}
\description{
Note that by default, the function tests whether input is a name in standard form.
}
\examples{

 is_region_name(c("Kainuu", "Kainuun maakunta"))
 is_region_name(c("Kainuu", "Kainuun maakunta"), allow_nonstandard_names = TRUE)

}
