% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_recode_regions.R
\name{join_abolished_mun}
\alias{join_abolished_mun}
\title{Join abolished municipalities}
\usage{
join_abolished_mun(x)
}
\arguments{
\item{x, }{a character vector of municipality codes.}
}
\value{

}
\description{
Takes a vector of municipality codes with KU-prefixes and transforms the codes of
the abolished municipalities to the codes of the municipality to which they have
joined.
}
\examples{

  v <- c("KU414", "KU609", "KU429", "KU273")
  join_abolished_mun(v)

}
