% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_with_key.R
\name{match_col}
\alias{match_col}
\title{Match columns to list elements}
\usage{
match_col(x, df, x_name = deparse(substitute(x)), by = "names")
}
\arguments{
\item{x}{vector, factor or data.frame}

\item{df}{data.frame where matcheds are found}
}
\value{
list with an element for each column in x. Each element contains all
columns that where matched in df.
}
\description{
Find columns in list that match vector by values or by names.
}
\examples{

v <- c("a", "b", "a", "c", "b", "b")
f <- factor(v)
df <- data.frame(var1 = v, y = rnorm(6))

key <- data.frame(var1 = letters[1:4],
                  var2 = c("first letter",
                           "second letter",
                           "third letter",
                           "fourth letter"))

 match_col(v, key, by = "values")
 var1 <- v
 match_col(var1, key)
 match_col(f, key, by = "values")
 match_col(df, key)
 match_col(df, key, by = "values")

key <- data.frame(var1 = letters[1:4],
                  var3 = letters[4:1],
                 var2 = c("first letter",
                          "second letter",
                          "third letter",
                          "fourth letter"))
 match_col(v, key, by = "values")
 match_col(df, key, by = "names")
 match_col(df, key, by = "values")




}
