% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_codes_and_names.R
\name{name_to_prefix}
\alias{name_to_prefix}
\title{Change region names to prefixes}
\usage{
name_to_prefix(name, pass_unknown = FALSE)
}
\arguments{
\item{name}{region level name}

\item{pass_unknown}{If TRUE pass unknown names as names.}
}
\value{
region prefix
}
\description{
Change region names to prefixes
}
\examples{

 name_to_prefix("seutukunta")

}
