% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_names_to_codes.R
\name{names_to_codes}
\alias{names_to_codes}
\title{Changes region codes to region names}
\usage{
names_to_codes(x, col = NULL)
}
\arguments{
\item{data}{data.frame, the input data that contains a variable of region codes.}

\item{region}{character, the name of the variable of region codes.}

\item{year}{integer, the year of the applied classification key.}
}
\value{
data.frame
}
\description{
A wrapper that uses the statficlassifications::recode-function.
}
\examples{

  v <- c("KOKO MAA", "Humppila", "Ålands skärgård", "Satakunta")
  names_to_codes(v)
  v <- c("KOKO MAA", "Humppila", "Forssa","Ålands skärgård", "Satakunta")
  names_to_codes(v)
  f <- factor(c("KOKO MAA", "Humppila","Ålands skärgård", "Satakunta"))
  names_to_codes(f)
  df <- data.frame(kunta_name = c("KOKO MAA", "Humppila", "Ålands skärgård", "Satakunta"),
                   values = rnorm(4))
  names_to_codes(df, "kunta_name")


}
