% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_names_to_codes.R
\name{names_to_codes}
\alias{names_to_codes}
\title{Changes region names to region codes}
\usage{
names_to_codes(x, year = NULL, offline = TRUE, region_level = NULL)
}
\arguments{
\item{year}{integer, the year of the applied classification key.}

\item{data}{data.frame, the input data that contains a variable of region codes.}

\item{region}{character, the name of the variable of region codes.}
}
\value{
data.frame
}
\description{
A wrapper that uses the statficlassifications::recode-function.
}
\examples{

  v <- c("KOKO MAA", "Humppila", "Ålands skärgård", "Satakunta")
  names_to_codes(v)
  names_to_codes("Kajaani", region_level = "seutukunta")
  f <- factor(c("KOKO MAA", "Humppila","Ålands skärgård", "Satakunta"))
  names_to_codes(f)

}
