% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_deal_with_region_codes_and_names.R
\name{names_to_codes}
\alias{names_to_codes}
\alias{names_to_codes_vct}
\alias{names_to_codes_fct}
\title{Change region names to region codes}
\usage{
names_to_codes(x, year = NULL, offline = TRUE, region_level = NULL)

names_to_codes_vct(x, year = NULL, offline = TRUE, region_level = NULL)

names_to_codes_fct(x, year = NULL, offline = TRUE, region_level = NULL)
}
\arguments{
\item{year}{integer, the year of the applied classification key.}

\item{offline}{logical, whether works offline with package data. Defaults to TRUE.}

\item{data}{data.frame, the input data that contains a variable of region codes.}

\item{region}{character, the name of the variable of region codes.}
}
\value{
data.frame
}
\description{
A wrapper that uses the statficlassifications::recode-function.
}
\section{Functions}{
\itemize{
\item \code{names_to_codes_vct}: Change region names to region codes

For internal use.

\item \code{names_to_codes_fct}: #' Recode name to codes in factors

For internal use.
}}

\examples{

  v <- c("KOKO MAA", "Humppila", "Ålands skärgård", "Satakunta")
  names_to_codes(v)
  names_to_codes("Kajaani", region_level = "seutukunta")
  names_to_codes(c("KOKO MAA", "Kajaani"), region_level = "seutukunta")
  f <- factor(c("KOKO MAA", "Humppila","Ålands skärgård", "Satakunta"))
  names_to_codes(f)

}
