% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_names_to_codes.R
\name{prefix_to_name}
\alias{prefix_to_name}
\title{Change prefixes to names and vice versa}
\usage{
prefix_to_name(prefix)
}
\arguments{
\item{prefix}{}
}
\value{

}
\description{
Change prefixes to names and vice versa
}
