% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_codes_and_names.R
\name{prefix_to_name}
\alias{prefix_to_name}
\title{Change region prefixes to names}
\usage{
prefix_to_name(prefix, pass_unknown = FALSE)
}
\arguments{
\item{prefix}{region level code prefix}

\item{pass_unknown}{If TRUE pass unknown prefix as prefix. Not in use}
}
\value{
region name
}
\description{
Change region prefixes to names
}
\examples{

 prefix_to_name("SK")

}
