% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_recoding_regions.R
\name{recode_region}
\alias{recode_region}
\title{Recode regional variables}
\usage{
recode_region(x, to, from = NULL, year = NULL, offline = TRUE)
}
\arguments{
\item{x}{data.frame, the input vector that contains the regional variable.}

\item{to}{character, the desired target in the classification key}

\item{from}{character, the original classification key}

\item{year}{character or numeric, the year of the}

\item{offline, }{logical, whether works offline with package data. Defaults to TRUE.}
}
\value{
data.frame
}
\description{
Recodes regional variables by e.g. changing kunnat to seutukunnat or changing names to codes
and vice versa. A wrapper for statfi_recode function that uses the 'get_regionkey'-function
to produce the key.
}
