% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionkey.R
\name{regionkey}
\alias{regionkey}
\title{Region classification keys}
\usage{
regionkey(region = "kunta", codes = FALSE, year = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{region}{character, the smallest region desired in the resulting classification key.}

\item{codes}{logical, whether the classification key contains region names or region codes. Defaults to FALSE.}

\item{year}{character or numerical, the year of desired classification key. Defaults to current year.}
}
\value{
data.frame Returns a classification key as a data.frame.
}
\description{
Imports NUTS-region classification keys from Statistics Finland. A Wrapper.
}
\examples{

}
