% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionkey.R
\docType{data}
\name{regionkey}
\alias{regionkey}
\title{Regionkey.}
\format{
A data frame with 310 rows and 10 variables:
\describe{
\item{}{}
\item{}{}
...
}
}
\source{
\url{}
}
\usage{
regionkey
}
\description{
Regionkey.
}
\keyword{datasets}
