% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api.R
\name{search_classifications}
\alias{search_classifications}
\title{Search for classifications}
\usage{
search_classifications(..., as_localId = FALSE)
}
\arguments{
\item{...}{character, search words.}

\item{as_localId}{logical, whether returns the localID of the found table. Defaults to \code{FALSE}.}
}
\description{
Searches and browses available classifications in the Classifications
classification service
}
\examples{

  search_classifications("ammatti")
  search_classifications("ammatti", year = 2021, as_localId = TRUE)

}
