% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_api_search.R
\name{search_keys}
\alias{search_keys}
\title{Search for classification keys}
\usage{
search_keys(
  ...,
  search_source = FALSE,
  search_target = FALSE,
  source = NULL,
  target = NULL,
  year = NULL,
  as_localId = FALSE
)
}
\arguments{
\item{...}{character, search words.}

\item{search_source}{logical, whether search only among the sources.}

\item{search_target}{logical, whether search only among the targets.}

\item{year}{character or numerical, search for specific years.}

\item{source_searchterm}{character, a search word to use in search only among the sources.}

\item{target_searchterm}{character, a search word to use in search only among the targets.}

\item{as_localID}{logical, whether returns the localID of the found table. Defaults to FALSE.}
}
\value{
character vector
}
\description{
A function to search and browse available classification keys / correspondence tables.
}
