% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_keys.R
\name{search_keys}
\alias{search_keys}
\title{Search for classification keys}
\usage{
search_keys(
  ...,
  search_source = FALSE,
  search_target = FALSE,
  searchword_source = NULL,
  searchword_target = NULL,
  year = NULL,
  as_localID = FALSE
)
}
\arguments{
\item{...}{character, the search words.}

\item{search_source}{logical, whether search only among the sources.}

\item{search_target}{logical, whether search only among the targets.}

\item{year}{character or numerical, search for specific years.}

\item{as_localID}{logical, whether returns the localID of the found table. Defaults to FALSE.}

\item{return}{character vector}
}
\value{

}
\description{
A function to search and browse Statistics Finland Classifications correspondence tables.
}
\examples{

# Browse all keys
   search_keys()

}
