% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_keys.R
\name{search_keys}
\alias{search_keys}
\title{Search for classification keys}
\usage{
search_keys(
  ...,
  search_source = FALSE,
  search_target = FALSE,
  searchword_source = NULL,
  searchword_target = NULL,
  year = NULL,
  as_localID = FALSE
)
}
\arguments{
\item{...}{character, search words.}

\item{search_source}{logical, whether search only among the sources.}

\item{search_target}{logical, whether search only among the targets.}

\item{searchword_source}{character, a search word to use in search only among the sources.}

\item{searchword_target}{character, a search word to use in search only among the targets.}

\item{year}{character or numerical, search for specific years.}

\item{as_localID}{logical, whether returns the localID of the found table. Defaults to FALSE.}
}
\value{
character vector
}
\description{
A function to search and browse Statistics Finland Classifications correspondence tables.
}
\examples{

# Browse all keys
   search_keys()

# Search for keys with search word "maakunta"
   search_keys("maakunta)

# Search for keys that have "maakunta" as source
   search_keys(search_source = "maakunta)

# Search for keys that map "kunta" to "maakunta"
   search_keys(searchword_source = "kunta", searchword_target = "maakunta")

# Search for keys that map "kunta" to "maakunta" for year 2016 and print as localID
  search_keys(searchword_source = "kunta", searchword_target = "maakunta",
              year = 2016, as_localID = TRUE)


}
