% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_names_to_codes.R, R/set_region_codes.R
\name{names_to_codes_df}
\alias{names_to_codes_df}
\alias{set_region_codes}
\title{Standardize region codes with prefixes}
\usage{
names_to_codes_df(x, col)

set_region_codes(x, col = NULL)
}
\arguments{
\item{x}{character vector of region codes}
}
\value{

}
\description{
Standardize region codes with prefixes
}
\section{Functions}{
\itemize{
\item \code{names_to_codes_df}: 
}}

\examples{


  v <- c("020", "047", "15", "133")
  set_region_codes(v)
  v <- c("Akaa" = "020", "Enontekiö" = "047", "Pohjanmaa" = "15", "Keuruu" =  "133")
  set_region_codes(v)
  v <- c("020" = "Akaa", "KU047" = "Enontekiö", "15" = "Pohjanmaa", "SK133" = "Keuruu")
  set_region_codes(names(v))
  f <- factor(c("020", "047", "15", "133"))
  set_region_codes(f)
  df <- data.frame(kunta_code = c("020", "047", "15", "133"), values = rnorm(4))
  set_region_codes(df, "kunta_code")

}
