% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_region_codes.R
\name{set_region_codes}
\alias{set_region_codes}
\title{Standardize region codes with prefixes}
\usage{
set_region_codes(x, region_level = NULL, col = NULL)
}
\arguments{
\item{x}{character vector of region codes}
}
\value{

}
\description{
Standardize region codes with prefixes
}
\examples{


  v <- c("020", "047", "15", "133")
  set_region_codes(v)
  v <- c("Akaa" = "020", "Enontekiö" = "047", "Pohjanmaa" = "15", "Keuruu" =  "133")
  set_region_codes(v)
  v <- c("020" = "Akaa", "KU047" = "Enontekiö", "15" = "Pohjanmaa", "SK133" = "Keuruu")
  set_region_codes(names(v))
  f <- factor(c("020", "047", "15", "133"))
  set_region_codes(f)

}
