% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_recoding_regions.R
\name{statfi_recode}
\alias{statfi_recode}
\title{Title}
\usage{
statfi_recode(x, ..., year = NULL, key = NULL, leave = FALSE)
}
\arguments{
\item{leave}{}
}
\value{

}
\description{
Title
}
