% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_recoding_regions.R
\name{statfi_recode}
\alias{statfi_recode}
\title{Recode classifications using key}
\usage{
statfi_recode(x, key, from = NULL, to = NULL)
}
\arguments{
\item{x}{vector of elements to be recoded}

\item{key}{the key to use in recoding}

\item{from}{source column in key for recoding}

\item{to}{target column in key for recoding}
}
\value{
vector or data.frame
}
\description{
Recode classifications using key
}
